Update workflow
Fix issue that pod/pods not there as offloading to autoscaler.
We override the values yaml with specifics for just the
GH action kind test with cli flags to ensure that its testing
the rollout correctly. Otherwise its only single pod, then the
autoscaling is building up to the min pod value but at the
point we run the deployment check its only a single pod so
means that the test is not really distributing over the
nodes via labels. This will fix this as its been present for
quite some time and was confusing me.